Remove unnecessary traits
