+++ dfpn-solver/src/impl_vec_table.rs
@@ -3,7 +3,6 @@ use crate::{Table, U};
 pub struct VecTable {
     table: Vec<Option<(U, U)>>,
     mask: usize,
-    len: usize,
 }
 
 impl VecTable {
@@ -11,7 +10,6 @@ impl VecTable {
         Self {
             table: vec![None; 1 << bits],
             mask: (1 << bits) - 1,
-            len: 0,
         }
     }
 }
@@ -27,15 +25,6 @@ impl Table for VecTable {
         self.table[(*key as usize) & self.mask].unwrap_or((1, 1))
     }
     fn put_in_hash(&mut self, key: u64, value: (U, U)) {
-        if self.table[(key as usize) & self.mask].is_none() {
-            self.len += 1;
-        }
         self.table[(key as usize) & self.mask] = Some(value);
-    }
-    fn len(&self) -> usize {
-        self.len
-    }
-    fn is_empty(&self) -> bool {
-        self.len == 0
     }
 }
